requirements.txt
discord.py
requests
beautifulsoup4
python-dotenv
apscheduler
httpx
lxml
openai>=1.0.0
